Move get_followees_for_user
