+++ discovery-provider/src/queries/queries.py
@@ -1,5 +1,6 @@
 import logging # pylint: disable=C0302
 import datetime
+from src.queries.get_followees_for_user import get_followees_for_user
 from src.queries.get_followers_for_user import get_followers_for_user
 from src.queries.get_tracks_including_unlisted import get_tracks_including_unlisted
 import sqlalchemy
@@ -33,6 +34,7 @@ from src.queries.get_follow_intersection_users import get_follow_intersection_us
 from src.queries.get_track_repost_intersection_users import get_track_repost_intersection_users
 from src.queries.get_playlist_repost_intersection_users import get_playlist_repost_intersection_users
 from src.queries.get_followers_for_user import get_followers_for_user
+from src.queries.get_followees_for_user import get_followees_for_user
 
 
 logger = logging.getLogger(__name__)
@@ -152,68 +154,8 @@ def get_followers_for_user_route(followee_user_id):
 
 # Get paginated users that are followed by provided follower_user_id, sorted by their follower count descending.
 @bp.route("/users/followees/<int:follower_user_id>", methods=("GET",))
-def get_followees_for_user(follower_user_id):
-    users = []
-    db = get_db_read_replica()
-    with db.scoped_session() as session:
-        # correlated subquery sqlalchemy code:
-        # https://groups.google.com/forum/#!topic/sqlalchemy/WLIy8jxD7qg
-        inner_follow = aliased(Follow)
-        outer_follow = aliased(Follow)
-
-        # subquery to get a user's follower count
-        inner_select = (
-            session.query(
-                func.count(inner_follow.followee_user_id)
-            )
-            .filter(
-                inner_follow.followee_user_id == outer_follow.followee_user_id,
-                inner_follow.is_current == True,
-                inner_follow.is_delete == False
-            )
-            .correlate(outer_follow)
-        )
-
-        # get all users followed by input user, sorted by their follower count desc
-        outer_select = (
-            session.query(
-                outer_follow.followee_user_id,
-                inner_select.as_scalar().label(response_name_constants.follower_count)
-            )
-            .filter(
-                outer_follow.follower_user_id == follower_user_id,
-                outer_follow.is_current == True,
-                outer_follow.is_delete == False
-            )
-            .group_by(outer_follow.followee_user_id)
-            .order_by(desc(response_name_constants.follower_count))
-        )
-        followee_user_ids_by_follower_count = paginate_query(outer_select).all()
-
-        user_ids = [user_id for (user_id, follower_count) in followee_user_ids_by_follower_count]
-
-        # get all users for above user_ids
-        users = (
-            session.query(User)
-            .filter(
-                User.is_current == True,
-                User.user_id.in_(user_ids)
-            )
-            .all()
-        )
-        users = helpers.query_result_to_list(users)
-
-        current_user_id = get_current_user_id(required=False)
-
-        # bundle peripheral info into user results
-        users = populate_user_metadata(session, user_ids, users, current_user_id)
-
-        # order by follower_count desc
-        users.sort(
-            key=lambda user: user[response_name_constants.follower_count],
-            reverse=True
-        )
-
+def get_followees_for_user_route(follower_user_id):
+    users = get_followees_for_user(follower_user_id)
     return api_helpers.success_response(users)
